Add files via upload
A template addon structure where addon name and "global variable" are called "template" & "TEMPLATE".
A lot of stuff to customize (sound, vgui images, language, ...).
A lot of exemple are available in files as comment.

This template is totaly working as a "stand alone". When you download it you can test it really quick in game, you have a entity called "Debug Item" and weapon called "Pickaxe". When you hit "Debug Item" with this pickaxe it play a sound and as a comment you already have a "spawning" logic for a entity such as a "rock piece" or whatever.